ai_service/requirements.txt
yfinance
bs4
textblob
uvicorn
fastapi
python-dotenv
pydantic
nsepython
pandas
lxml
html5lib
beautifulsoup4
selenium
webdriver-manager